requirements.txt
fastapi
uvicorn
httpx
slowapi
pyyaml
jinja2
python-multipart
